Fix case renaming.
For some reason case-renaming in git messed something up.
