+++ src/renderer/containers/automator/Automator.tsx
