+++ src/renderer/containers/automator/EventActions.tsx
