+++ src/renderer/containers/automator/StatusBar.tsx
