+++ src/renderer/containers/recorder/OBSStatusBar.tsx
