+++ src/renderer/containers/settings/ComboFinder.tsx
